Autenticación y registro de usuarios

+++ API/Routes/salesRoute.js
@@ -1,10 +1,19 @@
 const { Router } = require("express");
-const cors = require("cors");
+const {
+    checkRequiredPermissions,
+    validateAccessToken} = require("../Auth0/auth0.middleware.js");
+  const {
+  
+  } = require("../Auth0/auth0.permissions.js");
 const { getAllSales, createSale, getSaleById, updateSale, deleteSale } = require('../Controllers/sales.js')
 
 saleRouter = Router();
 
-saleRouter.get("/sale", cors(), async (req, res) => {
+saleRouter.get(
+    "/sale",
+    validateAccessToken,
+    checkRequiredPermissions([]),
+    async (req, res) => {
     //Si no hay productos en la BD, devuelve un arreglo vacío. NO es un error...
     try {
       const allSales = await getAllSales();
@@ -14,7 +23,11 @@ saleRouter.get("/sale", cors(), async (req, res) => {
     };
   });
   
-saleRouter.post("/sale", cors(), async (req, res) => {
+saleRouter.post(
+    "/sale",
+    validateAccessToken,
+    checkRequiredPermissions([]),
+    async (req, res) => {
   //Si algún dato no es válido o falta, se lanzan los errores correspondientes. Faltan las funciones validadoras.
     try {
         const { status, claim, products, user, location, paymentMethod, trackingCode, subtotal, shippingCost, taxes, total } = req.body;
@@ -37,7 +50,11 @@ saleRouter.post("/sale", cors(), async (req, res) => {
     };
 });
 
-saleRouter.get("/sale/:id", cors(), async (req, res) => {
+saleRouter.get(
+    "/sale/:id",
+    validateAccessToken,
+    checkRequiredPermissions([]),
+    async (req, res) => {
     try {
         const { id } = req.params;
         const sale = await getSaleById(id);
@@ -47,7 +64,11 @@ saleRouter.get("/sale/:id", cors(), async (req, res) => {
     };
 });
 
-saleRouter.put("/sale/:id", cors(), async (req, res) => {
+saleRouter.put(
+    "/sale/:id",
+    validateAccessToken,
+    checkRequiredPermissions([]),
+    async (req, res) => {
     try {
         const { id } = req.params;
         const update = req.body; //Hay que preguntar a los del frontend si están usando Axios o similares...
@@ -59,7 +80,11 @@ saleRouter.put("/sale/:id", cors(), async (req, res) => {
     };
 });
 
-saleRouter.delete("/sale/:id", cors(), async (req, res) => {
+saleRouter.delete(
+    "/sale/:id",
+    validateAccessToken,
+    checkRequiredPermissions([]),
+    async (req, res) => {
     try {
         const { id } = req.params;
         await deleteSale(id);
